refactor(lens): the row-styling leaf is generic, so its name says so
Adjudicating the one direct contradiction the lens review escalated: lowy
argued the CSS builder is a pure leaf that belongs in @kolu/solid-pierre;
hickey argued a git-agnostic wrapper package must not carry a git concept.

Both were right about different halves. Parameterizing the declaration body
(done earlier in this review) already removed the git DECISION — what remained
was git VOCABULARY on a function that is now genuinely generic: paths + decl ->
CSS. So rename ignoredPathsCss to rowPathsCss and drop 'ignored' from the
module. What it knows is Pierre ANATOMY — that a row is addressed by
data-item-path, and how a path survives as a CSS attribute value — which is
exactly this package's business and not the client's.

The git-specific reasoning (why the overlay must not ride Pierre's gitStatus
channel, with the measured 47/68 false 'contains changes' directories) moves to
the client call site, where a future reader deciding 'why not just use
gitStatus?' is actually standing.
